requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
google-genai>=0.1.0
pandas>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0